Refactor resource file for German language
The modification refactors the resource file for the German language. It updates the structure of the "Usages" and "Responses" sections by changing the indentation and removing unnecessary quotation marks. The changes also include adding missing indentation in a few places. This refactoring improves the readability and maintainability of the resource file.