Refactor: Update group and version information
This commit renames the group ID in the `pom.xml` file from `de` to `de.relaxogames` and updates the version from `inDev-1.9.3` to `inDev-1.10.0`. The package structure in the project has also been updated to reflect this change. The update to the group ID and version will ensure that the project is properly identified and versioned.