Update author list and website URL in config files
This commit introduces changes in both bungee.yml and plugin.yml files. The author list in the bungee.yml file was updated to include a new author, jotrorox. In the plugin.yml file, the website URL was changed to use https protocol instead of http, and the 'loadbefore' keys were removed as they are not needed now.